Renamed alg2 variables.
