+++ gettsim/tests/test_data/test_dfs_alg2.csv
@@ -1,4 +1,4 @@
-"hh_id","tu_id","p_id","vorstand_tu","kind","alter","kaltmiete_m","heizkost_m","wohnfläche","bewohnt_eigentum","alleinerziehend","bruttolohn_m","ges_rente_m","kapital_eink_m","arbeitsl_geld_m","sonstig_eink_m","eink_selbstst_m","vermiet_eink_m","eink_st_m","soli_st_m","sozialv_beitr_m","kindergeld_m_hh","unterhaltsvors_m","elterngeld_m","vermögen_hh","jahr","sum_basis_arbeitsl_geld_2_eink_hh","sum_arbeitsl_geld_2_eink_hh","sum_arbeitsl_geld_2_eink","alleinerziehenden_mehrbedarf","regelsatz_m","kost_unterk_m","regelbedarf_m","_arbeitsl_geld_2_eink","eink_anr_frei","vermfreibetr","maxvermfb","cost_per_sqm","wohnfl_just","hhsize","unterhaltsvors_m_hh","arbeitsl_geld_2_brutto_eink_hh","nettoquote",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+"hh_id","tu_id","p_id","vorstand_tu","kind","alter","kaltmiete_m","heizkost_m","wohnfläche","bewohnt_eigentum","alleinerziehend","bruttolohn_m","ges_rente_m","kapital_eink_m","arbeitsl_geld_m","sonstig_eink_m","eink_selbstst_m","vermiet_eink_m","eink_st_m","soli_st_m","sozialv_beitr_m","kindergeld_m_hh","unterhaltsvors_m","elterngeld_m","vermögen_hh","jahr","sum_basis_arbeitsl_geld_2_eink_hh","eink_anr_arbeitsl_geld_2_hh","eink_anr_arbeitsl_geld_2","alleinerziehenden_mehrbedarf","regelsatz_m","kost_unterk_m","regelbedarf_m","_arbeitsl_geld_2_eink","eink_anr_frei","vermfreibetr","maxvermfb","cost_per_sqm","wohnfl_just","hhsize","unterhaltsvors_m_hh","arbeitsl_geld_2_brutto_eink_hh","nettoquote",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 1,1,1,"TRUE","FALSE",30,700,100,75,"FALSE","FALSE",1000,0,0,0,0,0,0,100,20,200,190,0,0,0,2016,630,440,400,0,965,750,1715,680,280,,,10,75,3,0,1200,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 1,1,2,"FALSE","FALSE",30,700,100,75,"FALSE","FALSE",0,0,0,0,0,0,0,0,0,0,190,0,0,0,2016,630,440,0,0,965,750,1715,0,0,,,10,75,3,0,1200,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 1,1,3,"FALSE","TRUE",2,700,100,75,"FALSE","FALSE",200,0,0,0,0,0,0,0,0,40,190,0,0,0,2016,630,440,40,0,965,750,1715,160,120,,,10,75,3,0,1200,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
